Separate FuriEventLoopTimer to its own set of files
