+++ furi/furi.h
@@ -5,6 +5,7 @@
 #include "core/check.h"
 #include "core/common_defines.h"
 #include "core/event_loop.h"
+#include "core/event_loop_timer.h"
 #include "core/event_flag.h"
 #include "core/kernel.h"
 #include "core/log.h"
